fix : Separate the helper functions
